pull gfsFile stuff into its own file, re-organize code for cleaner diff
