+++ mongofiles/mongofiles.go
@@ -14,7 +14,6 @@ import (
 	"github.com/mongodb/mongo-go-driver/bson/primitive"
 	"github.com/mongodb/mongo-go-driver/mongo/gridfs"
 	driverOptions "github.com/mongodb/mongo-go-driver/mongo/options"
-	"github.com/mongodb/mongo-go-driver/x/bsonx"
 	"github.com/mongodb/mongo-tools-common/bsonutil"
 	"github.com/mongodb/mongo-tools-common/db"
 	"github.com/mongodb/mongo-tools-common/json"
@@ -26,7 +25,6 @@ import (
 	"io"
 	"os"
 	"regexp"
-	"time"
 )
 
 // List of possible commands for mongofiles.
@@ -69,123 +67,61 @@ type MongoFiles struct {
 	bucket *gridfs.Bucket
 }
 
-// Struct representing a GridFS files collection document.
-type gfsFile struct {
-	Id         interface{}        `bson:"_id"`
-	Name       string             `bson:"filename"`
-	Length     int64              `bson:"length"`
-	Md5        string             `bson:"md5"`
-	UploadDate time.Time          `bson:"uploadDate"`
-	Metadata   gfsFileMetadata    `bson:"metadata"`
-
-	// Storage required for reading and writing GridFS files
-	mf         *MongoFiles
-	downStream *gridfs.DownloadStream
-	upStream   *gridfs.UploadStream
-}
-
-// Struct representing the metadata associated with a GridFS files collection document.
-type gfsFileMetadata struct {
-	ChunkSize   int                `bson:"chunkSize"`
-	ContentType string             `bson:"contentType,omitempty"`
-}
-
-// Write data to GridFS Upload Stream. If this file has not been written to before, this function will open a new stream that must be closed.
-// Note: the go driver buffers data until it hits a chunk size before writing to the database, so if the amount of data is written < chunkSize,
-// the actual write to the database will occur in Close.
-func (file *gfsFile) Write(p []byte) (int, error) {
-	if file.upStream == nil {
-		rawBSON, err := bson.Marshal(file.Metadata)
-		if err != nil {
-			return 0, fmt.Errorf("could not marshal metadata to BSON: %v", err)
-		}
-
-		doc, err := bsonx.ReadDoc(rawBSON)
-		if err != nil {
-			return 0, fmt.Errorf("could not read metadata to document: %v", err)
-		}
-
-		// TODO: remove this (GO-815)
-		objectId, ok := file.Id.(primitive.ObjectID)
-		if !ok {
-			return 0, fmt.Errorf("need to use objectid for _id")
-		}
-
-		stream, err := file.mf.bucket.OpenUploadStreamWithID(objectId, file.Name, &driverOptions.UploadOptions{Metadata: doc})
-		if err != nil {
-			return 0, err
-		}
-		file.upStream = stream
+// ValidateCommand ensures the arguments supplied are valid.
+func (mf *MongoFiles) ValidateCommand(args []string) error {
+	// make sure a command is specified and that we don't have
+	// too many arguments
+	if len(args) == 0 {
+		return fmt.Errorf("no command specified")
 	}
 
-	return file.upStream.Write(p)
-}
-
-// Reads data from GridFS download stream. If this file has not been read from before, this function will open a new stream that must be closed.
-func (file *gfsFile) Read(buf []byte) (int, error) {
-	if file.downStream == nil {
-		// TODO: remove this (GO-815)
-		objectId, ok := file.Id.(primitive.ObjectID)
-		if !ok {
-			return 0, fmt.Errorf("need to use objectid for _id")
+	switch args[0] {
+	case List:
+		if len(args) > 2 {
+			return fmt.Errorf("too many positional arguments")
 		}
-
-		stream, err := file.mf.bucket.OpenDownloadStream(objectId)
-		if err != nil {
-			return 0, fmt.Errorf("could not open download stream: %v", err)
+		if len(args) == 1 {
+			mf.FileName = ""
+		} else {
+			mf.FileName = args[1]
 		}
-		file.downStream = stream
+	case Search, Put, Get, Delete:
+		if len(args) > 2 {
+			return fmt.Errorf("too many positional arguments")
+		}
+		// also make sure the supporting argument isn't literally an
+		// empty string for example, mongofiles get ""
+		if len(args) == 1 || args[1] == "" {
+			return fmt.Errorf("'%v' argument missing", args[0])
+		}
+		mf.FileName = args[1]
+	case GetID, DeleteID:
+		if len(args) > 2 {
+			return fmt.Errorf("too many positional arguments")
+		}
+		if len(args) == 1 || args[1] == "" {
+			return fmt.Errorf("'%v' argument missing", args[0])
+		}
+		mf.Id = args[1]
+	case PutID:
+		if len(args) > 3 {
+			return fmt.Errorf("too many positional arguments")
+		}
+		if len(args) < 3 || args[1] == "" || args[2] == "" {
+			return fmt.Errorf("'%v' argument(s) missing", args[0])
+		}
+		mf.FileName = args[1]
+		mf.Id = args[2]
+	default:
+		return fmt.Errorf("'%v' is not a valid command", args[0])
 	}
 
-	return file.downStream.Read(buf)
-}
-
-// Deletes the corresponding GridFS file in the database and its chunks.
-// Note: this file must be closed if it had been written to before being deleted. Any download streams will be closed as part of this deletion.
-func (file *gfsFile) Delete() error {
-	if file.upStream != nil {
-		return fmt.Errorf("this file (%v) must be closed for writing before being deleted", file.Name)
+	if mf.StorageOptions.GridFSPrefix == "" {
+		return fmt.Errorf("--prefix can not be blank")
 	}
 
-	if err := file.Close(); err != nil {
-		return err
-	}
-
-	// TODO: remove this (GO-815)
-	objectId, ok := file.Id.(primitive.ObjectID)
-	if !ok {
-		return fmt.Errorf("must be objectid")
-	}
-
-	if err := file.mf.bucket.Delete(objectId); err != nil {
-		return fmt.Errorf("error while removing '%v' from GridFS: %v\n", file.Name, err)
-	}
-
+	mf.Command = args[0]
 	return nil
-}
-
-// Closes any opened Download or Upload streams.
-func (file *gfsFile) Close() error {
-	if file.downStream != nil {
-		if err := file.downStream.Close(); err != nil {
-			return fmt.Errorf("could not close download stream: %v", err)
-		}
-	}
-
-	if file.upStream != nil {
-		if err := file.upStream.Close(); err != nil {
-			return fmt.Errorf("could not close upload stream: %v", err)
-		}
-	}
-
-	return nil
-}
-
-// Closes the file, assigning any error returned to finalErr. Use this to catch errors in deferred closing.
-func (file *gfsFile) SafeClose(finalErr *error) {
-	if err := file.Close(); err != nil {
-		*finalErr = err
-	}
 }
 
 // Query GridFS for files and display the results.
@@ -216,6 +152,21 @@ func (mf *MongoFiles) getLocalFileName(gridFile *gfsFile) string {
 		}
 	}
 	return localFileName
+}
+
+// handleGet contains the logic for the 'get' and 'get_id' commands
+func (mf *MongoFiles) handleGet() error {
+	file, err := mf.getTargetGFSFile()
+	if err != nil {
+		return err
+	}
+	defer file.SafeClose(&err)
+
+	if err = mf.writeGFSFileToFile(file); err != nil {
+		return err
+	}
+
+	return err
 }
 
 // Gets all GridFS files that match the given query.
@@ -296,6 +247,73 @@ func (mf *MongoFiles) deleteAll(filename string) error {
 	return nil
 }
 
+// handleDeleteID contains the logic for the 'delete_id' command
+func (mf *MongoFiles) handleDeleteID() error {
+	file, err := mf.getTargetGFSFile()
+	if err != nil {
+		return err
+	}
+
+	if err := file.Delete(); err != nil {
+		return err
+	}
+	log.Logvf(log.Always, fmt.Sprintf("successfully deleted file with _id %v from GridFS\n", mf.Id))
+
+	return nil
+}
+
+// parse and convert input extended JSON _id.
+func (mf *MongoFiles) parseID() (interface{}, error) {
+	if mf.Id == "" {
+		return primitive.NewObjectID(), nil
+	}
+
+	var asJSON interface{}
+	if err := json.Unmarshal([]byte(mf.Id), &asJSON); err != nil {
+		return nil, fmt.Errorf("error parsing provided extJSON: %v", err)
+	}
+
+	// legacy extJSON parser
+	id, err := bsonutil.ConvertJSONValueToBSON(asJSON)
+	if err != nil {
+		return nil, fmt.Errorf("error converting extJSON vlaue to bson: %v", err)
+	}
+
+	// TODO: fix this (GO-815)
+	mgoId, ok := id.(mgobson.ObjectId)
+	if !ok {
+		return nil, fmt.Errorf("only use ObjectIds as input _id")
+	}
+	objectId, err := primitive.ObjectIDFromHex(mgoId.Hex())
+	if err != nil {
+		return nil, err
+	}
+
+	return objectId, nil
+}
+
+// writeGFSFileToFile writes a file from gridFS to stdout or the filesystem.
+func (mf *MongoFiles) writeGFSFileToFile(gridFile *gfsFile) (err error) {
+	localFileName := mf.getLocalFileName(gridFile)
+	var localFile io.WriteCloser
+	if localFileName == "-" {
+		localFile = os.Stdout
+	} else {
+		if localFile, err = os.Create(localFileName); err != nil {
+			return fmt.Errorf("error while opening local file '%v': %v\n", localFileName, err)
+		}
+		defer localFile.Close()
+		log.Logvf(log.DebugLow, "created local file '%v'", localFileName)
+	}
+
+	if _, err = io.Copy(localFile, gridFile); err != nil {
+		return fmt.Errorf("error while writing Data into local file '%v': %v\n", localFileName, err)
+	}
+
+	log.Logvf(log.Always, fmt.Sprintf("finished writing to %s\n", localFileName))
+	return nil
+}
+
 // Write the given GridFS file to the database. Will fail if file already exists and --replace flag turned off.
 func (mf *MongoFiles) put(id interface{}, name string) (bytesWritten int64, finalErr error) {
 	gridFile := gfsFile{Name: mf.FileName, Id: id, mf: mf}
@@ -336,112 +354,21 @@ func (mf *MongoFiles) put(id interface{}, name string) (bytesWritten int64, fina
 	return n, nil
 }
 
-// writeGFSFileToFile writes a file from gridFS to stdout or the filesystem.
-func (mf *MongoFiles) writeGFSFileToFile(gridFile *gfsFile) (err error) {
-	localFileName := mf.getLocalFileName(gridFile)
-	var localFile io.WriteCloser
-	if localFileName == "-" {
-		localFile = os.Stdout
-	} else {
-		if localFile, err = os.Create(localFileName); err != nil {
-			return fmt.Errorf("error while opening local file '%v': %v\n", localFileName, err)
-		}
-		defer localFile.Close()
-		log.Logvf(log.DebugLow, "created local file '%v'", localFileName)
-	}
-
-	if _, err = io.Copy(localFile, gridFile); err != nil {
-		return fmt.Errorf("error while writing Data into local file '%v': %v\n", localFileName, err)
-	}
-
-	log.Logvf(log.Always, fmt.Sprintf("finished writing to %s\n", localFileName))
-	return nil
-}
-
-// parse and convert input extended JSON _id.
-func (mf *MongoFiles) parseID() (interface{}, error) {
-	if mf.Id == "" {
-		return primitive.NewObjectID(), nil
-	}
-
-	var asJSON interface{}
-	if err := json.Unmarshal([]byte(mf.Id), &asJSON); err != nil {
-		return nil, fmt.Errorf("error parsing provided extJSON: %v", err)
-	}
-
-	// legacy extJSON parser
-	id, err := bsonutil.ConvertJSONValueToBSON(asJSON)
+// handlePut contains the logic for the 'put' and 'put_id' commands
+func (mf *MongoFiles) handlePut() error {
+	id, err := mf.parseID()
 	if err != nil {
-		return nil, fmt.Errorf("error converting extJSON vlaue to bson: %v", err)
+		return err
 	}
 
-	// TODO: fix this (GO-815)
-	mgoId, ok := id.(mgobson.ObjectId)
-	if !ok {
-		return nil, fmt.Errorf("only use ObjectIds as input _id")
-	}
-	objectId, err := primitive.ObjectIDFromHex(mgoId.Hex())
+	n, err := mf.put(id, mf.FileName)
 	if err != nil {
-		return nil, err
+		return err
 	}
 
-	return objectId, nil
-}
+	log.Logvf(log.DebugLow, "copied %v bytes to server", n)
+	log.Logvf(log.Always, fmt.Sprintf("added gridFile: %v\n", mf.FileName))
 
-// ValidateCommand ensures the arguments supplied are valid.
-func (mf *MongoFiles) ValidateCommand(args []string) error {
-	// make sure a command is specified and that we don't have
-	// too many arguments
-	if len(args) == 0 {
-		return fmt.Errorf("no command specified")
-	}
-
-	switch args[0] {
-	case List:
-		if len(args) > 2 {
-			return fmt.Errorf("too many positional arguments")
-		}
-		if len(args) == 1 {
-			mf.FileName = ""
-		} else {
-			mf.FileName = args[1]
-		}
-	case Search, Put, Get, Delete:
-		if len(args) > 2 {
-			return fmt.Errorf("too many positional arguments")
-		}
-		// also make sure the supporting argument isn't literally an
-		// empty string for example, mongofiles get ""
-		if len(args) == 1 || args[1] == "" {
-			return fmt.Errorf("'%v' argument missing", args[0])
-		}
-		mf.FileName = args[1]
-	case GetID, DeleteID:
-		if len(args) > 2 {
-			return fmt.Errorf("too many positional arguments")
-		}
-		if len(args) == 1 || args[1] == "" {
-			return fmt.Errorf("'%v' argument missing", args[0])
-		}
-		mf.Id = args[1]
-	case PutID:
-		if len(args) > 3 {
-			return fmt.Errorf("too many positional arguments")
-		}
-		if len(args) < 3 || args[1] == "" || args[2] == "" {
-			return fmt.Errorf("'%v' argument(s) missing", args[0])
-		}
-		mf.FileName = args[1]
-		mf.Id = args[2]
-	default:
-		return fmt.Errorf("'%v' is not a valid command", args[0])
-	}
-
-	if mf.StorageOptions.GridFSPrefix == "" {
-		return fmt.Errorf("--prefix can not be blank")
-	}
-
-	mf.Command = args[0]
 	return nil
 }
 
@@ -504,63 +431,26 @@ func (mf *MongoFiles) Run(displayHost bool) (output string, finalErr error) {
 			regex := bson.M{"$regex": "^" + regexp.QuoteMeta(mf.FileName)}
 			query = bson.M{"filename": regex}
 		}
-
 		output, err = mf.findAndDisplay(query)
-		if err != nil {
-			return "", err
-		}
 
 	case Search:
 		regex := bson.M{"$regex": mf.FileName}
 		query := bson.M{"filename": regex}
 
 		output, err = mf.findAndDisplay(query)
-		if err != nil {
-			return "", err
-		}
 
 	case Get, GetID:
-		file, err := mf.getTargetGFSFile()
-		if err != nil {
-			return "", err
-		}
-		defer file.SafeClose(&finalErr)
-
-		if err = mf.writeGFSFileToFile(file); err != nil {
-			return "", err
-		}
+		err = mf.handleGet()
 
 	case Put, PutID:
-		id, err := mf.parseID()
-		if err != nil {
-			return "", err
-		}
-
-		n, err := mf.put(id, mf.FileName)
-		if err != nil {
-			return "", err
-		}
-
-		log.Logvf(log.DebugLow, "copied %v bytes to server", n)
-		log.Logvf(log.Always, fmt.Sprintf("added gridFile: %v\n", mf.FileName))
+		err = mf.handlePut()
 
 	case DeleteID:
-		file, err := mf.getTargetGFSFile()
-		if err != nil {
-			return "", err
-		}
-
-		if err := file.Delete(); err != nil {
-			return "", err
-		}
-
-		log.Logvf(log.Always, fmt.Sprintf("successfully deleted file with _id %v from GridFS\n", mf.Id))
+		err = mf.handleDeleteID()
 
 	case Delete:
-		if err := mf.deleteAll(mf.FileName); err != nil {
-			return "", err
-		}
+		err = mf.deleteAll(mf.FileName)
 	}
 
-	return output, nil
+	return output, err
 }
